+++ client/src/actions/user_actions.js
@@ -1,13 +1,7 @@
 import { userConstants } from "../constants/user_constants";
 import { userService } from "../services/user_services";
 import { alertActions } from "./alert_actions";
-
-export const userActions = {
-  login,
-  logout,
-  register,
-  _delete,
-};
+import { apiUrl } from "../config";
 
 function login(email, password) {
   return async (dispatch, getState) => {
@@ -15,13 +9,11 @@ function login(email, password) {
     const csrf = getState().csrf.csrfToken;
     try {
       const current_user = await userService.login(email, password, csrf);
-      console.log('CURRENT USER', current_user)
+      console.log("CURRENT USER", current_user);
       dispatch(success(current_user));
-
     } catch (error) {
       dispatch(failure(error.toString()));
       dispatch(alertActions.error(error.toString()));
-      
     }
   };
 
@@ -29,7 +21,7 @@ function login(email, password) {
     return { type: userConstants.LOGIN_REQUEST, user };
   }
   function success(user) {
-    console.log('SUCCESS USE ACTIONS', user)
+    console.log("SUCCESS USE ACTIONS", user);
     return { type: userConstants.LOGIN_SUCCESS, user };
   }
   function failure(error) {
@@ -37,8 +29,23 @@ function login(email, password) {
   }
 }
 
-function logout() {
-  userService.logout();
+const logout = () => async (dispatch, getState) => {
+  const csrf = getState().csrf.csrfToken;
+  const requestOptions = {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+      "X-CSRFToken": csrf,
+    },
+    credentials: "include",
+  };
+  const response = await fetch(`${apiUrl}/session/logout`, requestOptions);
+  if (response.ok) {
+    dispatch(removeUser());
+  }
+};
+
+function removeUser() {
   return { type: userConstants.LOGOUT };
 }
 
@@ -47,14 +54,14 @@ function register(user) {
     dispatch(request(user));
 
     try {
-    const csrf = getState().csrf.csrfToken;
-    const new_user = await userService.register(user, csrf);
-    dispatch(success(new_user));
-    dispatch(alertActions.success("Registration successful"));
-    } catch (error){
-        dispatch(failure(error.toString()));
-        dispatch(alertActions.error(error.toString()));
-      }
+      const csrf = getState().csrf.csrfToken;
+      const newUser = await userService.register(user, csrf);
+      dispatch(success(newUser));
+      dispatch(alertActions.success("Registration successful"));
+    } catch (error) {
+      dispatch(failure(error.toString()));
+      dispatch(alertActions.error(error.toString()));
+    }
   };
 
   function request(user) {
@@ -68,24 +75,8 @@ function register(user) {
   }
 }
 
-// prefixed function name with underscore because delete is a reserved word in javascript
-function _delete(id) {
-  return (dispatch) => {
-    dispatch(request(id));
-
-    userService.delete(id).then(
-      (user) => dispatch(success(id)),
-      (error) => dispatch(failure(id, error.toString()))
-    );
-  };
-
-  function request(id) {
-    return { type: userConstants.DELETE_REQUEST, id };
-  }
-  function success(id) {
-    return { type: userConstants.DELETE_SUCCESS, id };
-  }
-  function failure(id, error) {
-    return { type: userConstants.DELETE_FAILURE, id, error };
-  }
-}
+export const userActions = {
+  login,
+  logout,
+  register,
+};
